Put config in separate module
